projects/sandbox/export/pyproject.toml
[tool.poetry]
name = "export"
version = "0.0.1"
description = "Export DeepClean for IaaS inference"
authors = ["Alec Gunny <alec.gunny@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
export-model = "export:export"

[tool.poetry.dependencies]
python = "^3.8,<3.10"

"deepclean.architectures" = {path = "../../../libs/architectures", develop = true}
"deepclean.logging" = {path = "../../../libs/logging", develop = true}
"hermes.typeo" = {path = "../../../gw-iaas/libs/hermes/hermes.typeo", develop = true}

[tool.poetry.dependencies."hermes.quiver"]
path = "../../../gw-iaas/libs/hermes/hermes.quiver"
develop = true
extras = ["torch", "tensorflow"]

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[[tool.poetry.source]]
name = "ngc"
url = "https://pypi.ngc.nvidia.com"

[build-system]
requires = ["poetry>=1.2"]
build-backend = "poetry.masonry.api"
